+++ src/pke/demo/baseandpowers.cpp
@@ -59,7 +59,7 @@ void TestPowersAndDecompose(CryptoContext<ILVectorArray2n> cc, ILVectorArray2n& 
 		ILVectorArray2n::Integer twoPow( ILVectorArray2n::Integer(2).Exp(i * nBits) );
 		vector<ILVectorArray2n::ILVectorType> scalars(eParms->GetParams().size());
 
-		for( int t = 0; t < eParms->GetParams().size(); t++ ) {
+		for( size_t t = 0; t < eParms->GetParams().size(); t++ ) {
 			ILVectorArray2n::Integer factor = twoPow % mods[t];
 			ILVectorArray2n::ILVectorType thisScalar(eParms->GetParams()[t], EVALUATION);
 			thisScalar = factor.ConvertToInt();
