+++ src/pke/lib/pubkeylp.h
@@ -1018,8 +1018,8 @@ namespace lbcrypto {
 				RationalCiphertext<Element> determinant;
 				xCovariance.Determinant(&determinant);
 
-				for (int row = 0; row < result->GetRows(); row++)
-					for (int col = 0; col < result->GetCols(); col++)
+				for (size_t row = 0; row < result->GetRows(); row++)
+					for (size_t col = 0; col < result->GetCols(); col++)
 						(*result)(row, col).SetDenominator(*determinant.GetNumerator());
 
 				return result;
