+++ src/pke/unittest/UnitTestENCRYPT.cpp
@@ -49,7 +49,7 @@ UnitTestEncryption(const CryptoContext<Element>& cc) {
 	auto ptm = cc.GetCryptoParameters()->GetPlaintextModulus().ConvertToInt();
 
 	vector<uint32_t> intvec;
-	for( int ii=0; ii<intSize; ii++)
+	for( size_t ii=0; ii<intSize; ii++)
 		intvec.push_back( rand() % ptm );
 	IntPlaintextEncoding plaintextInt(intvec);
